Stash selection fix
