+++ st_widgets/stateful/selectbox.py
@@ -8,7 +8,7 @@ from streamlit.delta_generator import DeltaGenerator
 from streamlit.runtime.state import WidgetCallback
 from streamlit.type_util import Key, OptionSequence, T
 
-from st_widgets.stateful import _on_change_factory
+from st_widgets.stateful.base import _on_change_factory
 
 logger = get_logger(__name__)
